+++ fercord_storage/CHANGELOG.md
@@ -4,6 +4,9 @@
 
 ## [Unreleased] - ReleaseDate
 
+- feat: prelude now includes a lot more types
+- chore: dependency updates
+
 ## [0.3.5] - 2024-12-02
 - chore: Update dependencies